Add `calc` language, an arithmetic expression to help understand Pratt parsing (#1)
The `calc` language is a watered down version of `monkey` where I didn't just copy / paste stuff from the book but typed stuff over and made some minor changes. The `calc` program is a mere arithmetic expression, just enough to get to grips with Pratt parsing.

A good next move would be to port this to a familiar language like C#